requirements.txt
# requirements.txt
aiogram>=3.6.0
python-dotenv>=1.0.1
gspread>=6.0.0
oauth2client>=4.1.3